pyproject.toml
[tool.poetry]
name = "peft-play"
version = "0.1.0"
description = ""
authors = ["Stephen Chen <stephen.chen@utexas.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
torch = "^2.5.1"
numpy = "^2.2.0"
matplotlib = "^3.10.0"
peft = "^0.14.0"
jupyter = "^1.1.1"
hydra-core = "^1.3.2"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
sentencepiece = "^0.2.0"
protobuf = "^5.29.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
